runtime: Refactor things into a new `runtime` module
An attempt to better organize this crate.
